# target/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"Qos Slicing REST API","description":"Qos Slicing application REST API"},"basePath":"/onos/qosslice-app","produces":["application/json"],"consumes":["application/json"],"tags":[{"name":"qossliceapp","description":"Query and programm QoS Slice"}],"paths":{"/qossliceapp/{qosSliceName}":{"get":{"tags":["qossliceapp"],"summary":"Gets QosSlice","description":"Returns a QosSlice by qossliceName.","parameters":[{"name":"qosSliceName","in":"path","type":"string","description":"qosslice name","required":true}],"produces":["application/json"],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/QosSlice"}},"default":{"description":"Unexpected error"}}}},"/qossliceapp":{"get":{"tags":["qossliceapp"],"summary":"Gets all QosSlices","description":"Returns array of all QosSlices in the system.","parameters":[],"produces":["application/json"],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/QosSlices"}},"default":{"description":"Unexpected error"}}},"post":{"tags":["qossliceapp"],"summary":"Creates new QosSlice","description":"Creates and installs a new QosSlice.<br>","parameters":[{"name":"stream","in":"body","schema":{"$ref":"#/definitions/QosSlicePost"},"description":"QosSlice JSON","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/qossliceapp/meters/{qossliceName}":{"post":{"tags":["qossliceapp"],"summary":"Adds a queues","description":"","parameters":[{"name":"qossliceName","in":"path","type":"string","description":"Qos Slice name","required":true},{"name":"stream","in":"body","schema":{"$ref":"#/definitions/MeterPost"},"required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}},"delete":{"tags":["qossliceapp"],"summary":"Removes a meters","description":"","parameters":[{"name":"qossliceName","in":"path","type":"string","description":"Qos Slice name","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/qossliceapp/{qossliceName}":{"delete":{"tags":["qossliceapp"],"summary":"Removes the specified QosSlice","description":"","parameters":[{"name":"qossliceName","in":"path","type":"string","description":"QosSlice name","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/qossliceapp/queues/{qossliceName}":{"post":{"tags":["qossliceapp"],"summary":"Adds a queues","description":"","parameters":[{"name":"qossliceName","in":"path","type":"string","description":"Qos Slice name","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}},"delete":{"tags":["qossliceapp"],"summary":"Removes a queues","description":"","parameters":[{"name":"qossliceName","in":"path","type":"string","description":"Qos Slice name","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}}},"definitions":{"QosSlices":{"type":"object","title":"qosslices","required":["vpls-name","qos-name"],"properties":{"vpls-name":{"type":"String","example":"VPLS1"},"qos-name":{"type":"String","example":"SLICE1"},"queue":{"type":"Boolean","example":true},"meter":{"type":"Boolean","example":true},"unit":{"type":"string","example":"KB_PER_SEC"},"bands":{"type":"array","xml":{"name":"bands","wrapped":true},"items":{"type":"object","title":"bands","required":["type","rate","burstSize","prec"],"properties":{"type":{"type":"string","example":"REMARK"},"rate":{"type":"integer","format":"int64","example":0},"burstSize":{"type":"integer","format":"int64","example":0},"prec":{"type":"integer","format":"int16","example":0}}}}}},"QosSlice":{"type":"object","title":"qosslice","required":["vpls-name","qos-name"],"properties":{"vpls-name":{"type":"String","example":"VPLS1"},"qos-name":{"type":"String","example":"SLICE1"},"queue":{"type":"Boolean","example":true},"meter":{"type":"Boolean","example":true},"unit":{"type":"string","example":"KB_PER_SEC"},"bands":{"type":"array","xml":{"name":"bands","wrapped":true},"items":{"type":"object","title":"bands","required":["type","rate","burstSize","prec"],"properties":{"type":{"type":"string","example":"REMARK"},"rate":{"type":"integer","format":"int64","example":0},"burstSize":{"type":"integer","format":"int64","example":0},"prec":{"type":"integer","format":"int16","example":0}}}}}},"QosSlicePost":{"type":"object","title":"qosslice","required":["vpls-name","qos-name"],"properties":{"vpls-name":{"type":"String","example":"VPLS1"},"qos-name":{"type":"String","example":"SLICE1"},"queue":{"type":"Boolean","example":true},"meter":{"type":"Boolean","example":true},"unit":{"type":"string","example":"KB_PER_SEC"},"bands":{"type":"array","xml":{"name":"bands","wrapped":true},"items":{"type":"object","title":"bands","required":["type","rate","burstSize","prec"],"properties":{"type":{"type":"string","example":"REMARK"},"rate":{"type":"integer","format":"int64","example":0},"burstSize":{"type":"integer","format":"int64","example":0},"prec":{"type":"integer","format":"int16","example":0}}}}}},"MeterPost":{"type":"object","title":"band","required":["type","rate","burstSize","prec"],"properties":{"type":{"type":"string","example":"REMARK"},"rate":{"type":"integer","format":"int64","example":0},"burstSize":{"type":"integer","format":"int64","example":0},"prec":{"type":"integer","format":"int16","example":0}}}}}